+++ core/server/api_container/server/startosis_engine/startosis_executor.go
@@ -6,6 +6,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/binding_constructors"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_instruction"
 	"github.com/kurtosis-tech/stacktrace"
+	"strings"
 	"sync"
 )
 
@@ -24,17 +25,21 @@ func NewStartosisExecutor() *StartosisExecutor {
 }
 
 // Execute executes the list of Kurtosis instructions against the Kurtosis backend
-// It returns a potential execution error if something went wrong.
+// It serializes each instruction that is executed and returned the list of serialized instruction as a result
 // It returns an error if something unexpected happens outside the execution of the script
-func (executor *StartosisExecutor) Execute(ctx context.Context, dryRun bool, instructions []kurtosis_instruction.KurtosisInstruction) ([]*kurtosis_core_rpc_api_bindings.SerializedKurtosisInstruction, error) {
+func (executor *StartosisExecutor) Execute(ctx context.Context, dryRun bool, instructions []kurtosis_instruction.KurtosisInstruction, outputBuffer *strings.Builder) ([]*kurtosis_core_rpc_api_bindings.SerializedKurtosisInstruction, error) {
 	executor.mutex.Lock()
 	defer executor.mutex.Unlock()
 
 	var serializedSuccessfullyExecutedInstructions []*kurtosis_core_rpc_api_bindings.SerializedKurtosisInstruction
 	for index, instruction := range instructions {
 		if !dryRun {
-			if err := instruction.Execute(ctx); err != nil {
+			instructionOutput, err := instruction.Execute(ctx)
+			if err != nil {
 				return serializedSuccessfullyExecutedInstructions, stacktrace.Propagate(err, "An error occurred executing instruction (number %d): \n%v", index+1, instruction.GetCanonicalInstruction())
+			}
+			if instructionOutput != nil {
+				outputBuffer.WriteString(*instructionOutput)
 			}
 		}
 		serializedSuccessfullyExecutedInstructions = append(serializedSuccessfullyExecutedInstructions,