+++ docs/changelog.md
@@ -3,6 +3,14 @@
 ### Features
 - Log that the module is being compressed & uploaded during `kurtosis exec`
 
+### Changes
+- `print()` is now a regular instructions like others, and it takes effect at execution time (used to be during interpretation)
+
+# 0.53.2
+### Features
+- Make facts referencable on `add_service`
+- Added a new log line for printing the `created enclave ID` just when this is created in `kurtosis exec` and `kurtosis module exec` commands
+
 # 0.53.1
 ### Features
 - Added random enclave ID generation in `EnclaveManager.CreateEnclave()` when an empty enclave ID is provided
@@ -20,6 +28,7 @@
 - Made `render_templates`, `upload_files`, `store_Files_from_service` accept `artifact_uuid` and
 return `artifact_uuid` during interpretation time
 - Moved `kurtosis startosis exec` to `kurtosis exec`
+- Added `import_module` startosis builtin to replace `load`. Load is now deprecated. It can still be used but it will log a warning. It will be entirely removed in a future PR
 
 ### Breaking Features
 - Moved `kurtosis startosis exec` to `kurtosis exec`